intermediate update: some clean-ups and hunting a nasty bug in CVWidget2D when used in CVCenter (lots of debugging-output in this commit)
- clean-up: consequently refer to \hi or \lo as 'slot' in methods
('key' should always be reserved for the widget's name within CVCenter)
- bughunting: when using 2D-widgets in CVCenter and the main GUI gets
closed and reopened OSC- and MIDI-responders do basically work but the
slot (hi or lo) that has been created later will not reflect the
current value of the CV properly (e.g. in the 2D-slider the cursor will
be hanging on the left side, only moving up and down not left or right
or the other way round - depends on which slot has been created first).
Had a look if my usage of OSCresponderNode resp. CCResponder caused the
issue but using OSCFunc instead didn't change anything...